+++ proto/proto.h
@@ -1,6 +1,6 @@
 /* proto.h
  *   by Trinity Quirk <tquirk@ymb.net>
- *   last updated 14 Apr 2018, 21:21:13 tquirk
+ *   last updated 16 Apr 2018, 07:33:13 tquirk
  *
  * Revision IX game server
  * Copyright (C) 2018  Trinity Annabelle Quirk
@@ -84,7 +84,7 @@ typedef struct basic_packet_tag
 {
     uint8_t type;
     uint8_t version;
-    uint64_t sequence;
+    uint32_t sequence;
 } __attribute__ ((__packed__))
 basic_packet;
 
@@ -95,7 +95,7 @@ typedef struct ack_packet_tag
 {
     uint8_t type;
     uint8_t version;        /* protocol version number */
-    uint64_t sequence;      /* timestamp / sequence number */
+    uint32_t sequence;      /* timestamp / sequence number */
     uint8_t request;        /* packet type of original request */
     uint64_t misc[4];       /* miscellaneous data */
 } __attribute__ ((__packed__))
@@ -105,7 +105,7 @@ typedef struct login_request_tag
 {
     uint8_t type;
     uint8_t version;        /* protocol version number */
-    uint64_t sequence;      /* timestamp / sequence number */
+    uint32_t sequence;      /* timestamp / sequence number */
     char username[64];
     char password[64];
     char charname[64];
@@ -116,7 +116,7 @@ typedef struct action_request_tag
 {
     uint8_t type;
     uint8_t version;        /* protocol version number */
-    uint64_t sequence;      /* timestamp / sequence number */
+    uint32_t sequence;      /* timestamp / sequence number */
     uint64_t object_id;
     uint16_t action_id;
     uint8_t power_level;
@@ -139,7 +139,7 @@ typedef struct position_update_tag
 {
     uint8_t type;
     uint8_t version;        /* protocol version number */
-    uint64_t sequence;      /* timestamp / sequence number */
+    uint32_t sequence;      /* timestamp / sequence number */
     uint64_t object_id;
     uint16_t frame_number;
     /* We may consider adding the sector vector back in here */
@@ -154,7 +154,7 @@ typedef struct server_notice_tag
 {
     uint8_t type;
     uint8_t version;        /* protocol version number */
-    uint64_t sequence;      /* timestamp / sequence number */
+    uint32_t sequence;      /* timestamp / sequence number */
     uint8_t ipproto;        /* 4 or 6 */
     char addr[INET6_ADDRSTRLEN];
     in_port_t port;
